% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_ggplotting_rep.R
\name{calEDscore}
\alias{calEDscore}
\title{calEDscore}
\usage{
calEDscore(data, nread, nreplicate)
}
\arguments{
\item{data}{melt curve dataset (in complete replicates)}

\item{nread}{number of reading channels or sample treatments}

\item{nreplicate}{number of replicates}
}
\value{
a dataframe
}
\description{
Internal function to calculate euclidean distance metrics
for CETSA melt curve data
}
\examples{
\dontrun{
}

}
\keyword{internal}
