% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_NADPH.R
\docType{data}
\name{data_NADPH}
\alias{data_NADPH}
\title{ITDR data of NADP/NADPH on K562 lysate.}
\format{A data frame with 24766 rows and 10 doses:
\describe{t
  \item{id}{Uniprot ID}
  \item{description}{Protein annotation}
  \item{condition}{Experimental treatment condition plus replicate number}
  ...
}}
\source{
Nayana Prabhu/Par Nordlund
}
\usage{
attach(data_NADPH)
}
\description{
A dataset containing the NADP/NADPH ITDR CETSA data on K562 cell lysate.
Search on Proteome Discover 2.0, Cleaned, Scaled, and Fitted.
}
\keyword{datasets}
