% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_twoUnknownDrugs.R
\docType{data}
\name{data_twoUnknownDrugs}
\alias{data_twoUnknownDrugs}
\title{ITDR data of two unknown drugs on K562 lysate.}
\format{A data frame with 18291 rows and 10 doses:
\describe{
  \item{id}{Uniprot ID}
  \item{description}{Protein annotation}
  \item{condition}{Experimental treatment condition plus replicate number}
  ...
}}
\source{
Anderson Ramos/Par Nordlund
}
\usage{
attach(data_twoUnknownDrugs)
}
\description{
A dataset containing two unknown drugs ITDR CETSA data on K562 cell lysate.
Search on Proteome Discover 2.1, Cleaned, Scaled, and Fitted.
}
\keyword{datasets}
