% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_graphs.R
\name{external_graphs}
\alias{external_graphs}
\title{external_graphs}
\usage{
external_graphs(ext = TRUE)
}
\arguments{
\item{ext}{whether to switch on the external graph device}
}
\value{
a command
}
\description{
Internal function to change rstudio default graph device
}
\examples{
\dontrun{
}

}
\keyword{internal}
