% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getName.R
\name{getGeneName}
\alias{getGeneName}
\title{getName}
\usage{
getGeneName(x)
}
\arguments{
\item{x}{Uniprot description input}

\item{pfdatabase}{whether the input is from P.falciparum database}
}
\value{
character
}
\description{
Functions to extract Gene/Protein/Both names from the Uniprot description
}
\examples{
\dontrun{
}
}
\keyword{internal}
