% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITDR_NADPH.R
\docType{data}
\name{ITDR_NADPH}
\alias{ITDR_NADPH}
\title{ITDR data of NADP/NADPH on K562 lysate without reducing TCEP.}
\format{A data frame with 17524 rows and 10 doses:
\describe{t
  \item{id}{Uniprot ID}
  \item{description}{Protein annotation}
  \item{condition}{Experimental treatment condition plus replicate number}
  ...
}}
\source{
Nayana Prabhu/Par Nordlund
}
\usage{
attach(ITDR_NADPH)
}
\description{
A dataset containing the NADP/NADPH ITDR CETSA data on K562 cell lysate.
Search on Proteome Discover 2.1, Cleaned, Scaled, and Fitted.
}
\keyword{datasets}
