% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_2D_average.R
\name{ms_2D_average}
\alias{ms_2D_average}
\title{ms_2D_average}
\usage{
ms_2D_average(data)
}
\arguments{
\item{data}{dataset after calculating the relative protein abundance differences}
}
\value{
a dataframe
}
\description{
Function to calculate the averaged signals from the 2D result
}
\examples{
\dontrun{
  MOLM <- ms_2D_average(MOLM)
}


}
