% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_2D_barplotting.R
\name{ms_2D_barplotting}
\alias{ms_2D_barplotting}
\title{ms_2D_barplotting}
\usage{
ms_2D_barplotting(data, treatmentlevel = NULL, setlevel = NULL,
  printBothName = TRUE, printGeneName = FALSE, pfdatabase = FALSE,
  witherrorbar = TRUE, colorpanel = c("gray", "blue", "orange"),
  layout = NULL, external = TRUE, log2scale = TRUE,
  toplabel = "2D-CETSA bar plotting", leftlabel = "", bottomlabel = "",
  pdfname = "bar_ggplotting.pdf", pdfheight = 12, pdfwidth = 12)
}
\arguments{
\item{data}{dataset after ms_2D_caldiff to plot}

\item{treatmentlevel}{a vector of treatment labels, such as c("DMSO","TNFa","AT26533")
the order determines the arrangement, so in this case DMSO group would be the first group}

\item{setlevel}{a vector of set information if any, such as c("M13","M16")}

\item{colorpanel}{a vector of customizable color scheme provided by the user}

\item{layout}{a vector indicating the panel layout for multi-panel plots
per page, default value is c(2,3) for set containing data, otherwise c(4,3)}

\item{toplabel}{textual label at the top part of the page}

\item{leftlabel}{textual label at the left side of the page}

\item{bottomlabel}{textual label at the bottom part of the page}

\item{pdfheight}{a number indicate the height of pdf file, default value 12}

\item{pdfwidth}{a number indicate the width of pdf file, default value 12}
}
\value{
a list of ggplot2 object
}
\description{
Function to generate pdf files with multipanel bar plots for 2D-CETSA data
}
\examples{
\dontrun{
  ms_2D_barplotting(MOLM, treatmentlevel=c("DMSO","TNFa","AT26533"), setlevel=c("M13","M16"))
}


}
