% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_2D_DE.R
\name{ms_2D_DE}
\alias{ms_2D_DE}
\title{ms_2D_DE}
\usage{
ms_2D_DE(data, set = NULL, treatment = NULL, cvthreshold = 0.1,
  basetemp = "37C", expressionchange_nMAD = 2.5,
  expressionchange_cutoff = NULL, densityformat = FALSE, binwidth = NULL,
  xlimit = NULL)
}
\arguments{
\item{data}{dataset after ms_2D_caldiff() function}

\item{set}{a single character to indicate the sample name}

\item{treatment}{a single character to indicate the sample name}

\item{cvthreshold}{the CV threshold value for subsetting reproducible
measurements, default value is 0.1}

\item{basetemp}{the character indicating the baseline temperature for
expression levels, default value is 37C}

\item{expressionchange_nMAD}{the number of MADs to set the significance
cutoff on protein expression level change, default value is 2.5}

\item{expressionchange_cutoff}{a two numeric element vector to indicate
the significance cutoff on protein expression level change}

\item{densityformat}{whether plot the histogram in density plot format}

\item{binwidth}{the width of bin used in histogram plot}

\item{xlimit}{a two numeric element vector to indicate the limit of x-axis}
}
\value{
a dataframe
}
\description{
Function to have an overview of the 2D CETSA dataset, regarding to
expression level changes only, to select out the differentially expressed ones
}
\examples{
\dontrun{
  MOLM <- ms_2D_DE(MOLM, set="M13", treatment="AT26533")
}

}
