% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_2D_diffExp.R
\name{ms_2D_diffExp}
\alias{ms_2D_diffExp}
\title{ms_2D_diffExp}
\usage{
ms_2D_diffExp(data, set = NULL, basetemp = "37C", contrast = NULL,
  logFC_threshold = 0.2, adjp_threshold = 0.01, labelnodes = TRUE,
  xlimit = NULL, ylimit = NULL, returneset = FALSE)
}
\arguments{
\item{data}{dataset after ms_2D_normalization function, readings in log2 format}

\item{set}{a single character to indicate the sample name}

\item{basetemp}{the character indicating the baseline temperature for
expression levels, default value is 37C}

\item{contrast}{a character to indicate the contrasting treatment conditions}

\item{logFC_threshold}{the threshold value for log fold changes, default set at 0.2}

\item{adjp_threshold}{the threshold value for adjusted p values, default set at 0.01}

\item{labelnodes}{whether to label the proteins with significant differential expression}

\item{xlimit}{a two numeric element vector to indicate the limit of x-axis}

\item{ylimit}{a two numeric element vector to indicate the limit of y-axis}
}
\value{
a dataframe
}
\description{
Function to have an overview of the 2D CETSA dataset, regarding to
expression level changes only, to select out the differentially expressed proteins
}
\examples{
\dontrun{
  MOLM <- ms_2D_diffExp(MOLM, set="M13", contrast="TNFa-DMSO")
}

}
