% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_2D_rawread.R
\name{ms_2D_rawread}
\alias{ms_2D_rawread}
\title{ms_2D_rawread}
\usage{
ms_2D_rawread(filevector, fchoose = FALSE, treatment = NULL, nread = 10,
  abdread = TRUE, PD21 = TRUE, noratiodata = TRUE, refchannel = "126",
  channels = c("126", "127N", "127C", "128N", "128C", "129N", "129C", "130N",
  "130C", "131"))
}
\arguments{
\item{filevector}{a file name or a vector of filenems to import}

\item{fchoose}{whether to choose file interactively, default set to FALSE}

\item{treatment}{a vector of treatment names applied to CETSA samples, in the same order as channels,
in a typical 2D setting, the treatment name should contain both replicate and treatment information,
preferably concatenated with an underline}

\item{nread}{number of reading channels, should match the number of channels used, default value 10}

\item{abdread}{whether to read in protein abundance data, default set to TRUE}

\item{PD21}{whether the data is searched using Proteome Discoverer 2.1, default set to TRUE}

\item{refchannel}{names of reference channel used in Proteome Discoverer search, default value 126}

\item{channels}{names of the read-in channels, default value c("126","127N","127C","128N","128C","129N","129C","130N","130C","131")}
}
\value{
a dataframe
}
\description{
Function to parse and read in CETSA data from tab delimited files exported from Proteome Discoverer
}
\examples{
\dontrun{
 MOLM13 <- ms_2D_rawread(c("M13_37C_frac_Proteins.txt","M13_52C_frac_Proteins.txt","M13_58C_frac_Proteins.txt"), nread=10,
 treatment=("B1_DMSO","B1_TNFa","B1_AT26533","B2_DMSO","B2_TNFa","B2_AT26533","B3_DMSO","B3_TNFa","B3_AT26533","Mix"))
}



}
\seealso{
\code{\link{ms_ITTR_rawread}} for time response data

\code{\link{ms_IDTR_rawread}} for dose response data
}
