% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_2D_rearrange.R
\name{ms_2D_rearrange}
\alias{ms_2D_rearrange}
\title{ms_2D_rearrange}
\usage{
ms_2D_rearrange(data, nread = 9, repthreshold = 0.75, averagecount = TRUE,
  countthreshold = 2)
}
\arguments{
\item{data}{dataset to transform into a wide format}

\item{nread}{number of reading channels, default value 9 in 2D-CETSA scheme}

\item{repthreshold}{the minimal percentage threshold of protein being sampled
from multiple runs, default value is 0.75}

\item{averagecount}{whether to average the supporting PSM/peptide/count numbers,
default set to TRUE}

\item{countthreshold}{the minimal threshold number of associated abundance
count of proteins, default value is 2}
}
\value{
a dataframe
}
\description{
Function to rearrange 2D-CETSA data into a wide one-unique-protein-per-row format
}
\examples{
\dontrun{
 MOLM <- ms_rearrange(MOLM, nread=9, countthreshold = 3)
}


}
