% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_2D_reproducible.R
\name{ms_2D_reproducible}
\alias{ms_2D_reproducible}
\title{ms_2D_reproducible}
\usage{
ms_2D_reproducible(data, set = NULL, treatment = NULL, cvthreshold = 0.1,
  corrthreshold = 0.5, basetemp = "37C")
}
\arguments{
\item{data}{dataset after ms_2D_caldiff() function}

\item{set}{a single character to indicate the sample name}

\item{treatment}{a single character to indicate the sample name}

\item{cvthreshold}{the CV threshold value for subsetting reproducible
measurements, default value is 0.1}

\item{corrthreshold}{the Correlation threshold value for subsetting
reproducible measurements, default value is 0.5}

\item{basetemp}{the character indicating the baseline temperature for
expression levels, default value is 37C}
}
\value{
a dataframe
}
\description{
Function to retrieve a reproducible subset of the 2D CETSA dataset, regarding to
expression level changes and thermal shifts
}
\examples{
\dontrun{
  MOLM <- ms_2D_reproducible(MOLM)
}

}
