% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_2D_score.R
\name{ms_2D_score}
\alias{ms_2D_score}
\title{ms_2D_score}
\usage{
ms_2D_score(data, basetemp = "37C", allzscore = TRUE, fdrthreshold = 0.01,
  labelnodes = FALSE, labelcategory = c("CC", "NC", "CN"), xrange = NULL,
  yrange = NULL)
}
\arguments{
\item{data}{dataset after calculating the relative protein abundance differences}

\item{basetemp}{the character indicating the baseline temperature for
expression levels, default value is 37C}

\item{allzscore}{when calculating the z score, whether to use all the readings from
all the different treatment groups, default set to TRUE}

\item{fdrthreshold}{the significance level of global fdr, default value is 0.01}

\item{labelnodes}{whether to text-label the selected nodes, default set to FALSE}

\item{labelcategory}{the categories of nodes to label, default value is c("CC","NC","CN")}
}
\value{
a dataframe
}
\description{
Function to calculate the relative abundance score and thermal stability score
}
\examples{
\dontrun{
  MOLM <- ms_2D_score(MOLM)
}


}
