% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_clean_conditionrename.R
\name{ms_clean}
\alias{ms_clean}
\title{ms_clean}
\usage{
ms_clean(data, nread = 10, remkeratin = TRUE, remsinglecondprot = FALSE)
}
\arguments{
\item{data}{dataset to be cleaned up}

\item{nread}{number of reading channels or sample treatements, default value 10}

\item{remkeratin}{whether to remove Keratin protein, which is generally considered as a common contaminant, default set to TRUE}

\item{remsinglecondprot}{whether the orphan proteins that appear only in one of the datasets should be removed, default set to FALSE}
}
\value{
a dataframe after clean-up
}
\description{
Function to clean up read-in dataset, mainly to remove not-quantified entries
}
\examples{
\dontrun{
 ITDRdata_cleaned <- ms_clean(ITDRdata)
}


}
