% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_melt_innerplot.R
\name{ms_complex_melt_innerplot}
\alias{ms_complex_melt_innerplot}
\title{ms_complex_melt_innerplot}
\usage{
ms_complex_melt_innerplot(data, nread, topasone, printpvalue, colorpanel,
  bycondition, toplabel, leftlabel, bottomlabel, layout)
}
\arguments{
\item{data}{melt curve dataset with complex information to plot}

\item{nread}{number of reading channels or sample treatements}

\item{topasone}{whether the top plateau has to be fixed, i.e., 1.0}

\item{colorpanel}{a vector of customizable color scheme provided by the user}

\item{bycondition}{whether to color by condition}

\item{toplabel}{textual label at the top of the page}

\item{leftlabel}{textual label at the left side of the page}

\item{bottomlabel}{textual label at the bottom of the page}

\item{layout}{a vector indicating the panel layout for multi-panel plots per page}
}
\value{
a list of ggplot2 object
}
\description{
Internal function to generate ggplot objects for CETSA melt curve data grouped by complex
}
\examples{
\dontrun{
}


}
\keyword{internal}
