% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_composite.R
\name{ms_composite_ID_Protein}
\alias{ms_composite_ID_Protein}
\title{ms_composite_ID_Protein}
\usage{
ms_composite_ID_Protein(data, pfdatabase = FALSE)
}
\arguments{
\item{data}{a dataframe with id and description column}

\item{pfdatabase}{whether the input is from P.falciparum database}
}
\value{
a dataframe with a compositie id column
}
\description{
Functions to make a composite Uniprot ID/Protein name from the Uniprot description
}
\examples{
\dontrun{
}
}
