% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_clean_conditionrename.R
\name{ms_conditionrename}
\alias{ms_conditionrename}
\title{ms_conditionrename}
\usage{
ms_conditionrename(data, incondition = NULL, outcondition = NULL)
}
\arguments{
\item{data}{dataset of which the condition names to be customized}

\item{incondition}{a vector of current condition naming to be changed}

\item{outcondition}{a vector of new condition naming to be applied}
}
\value{
a dataframe after condition name customization
}
\description{
Function to customize the condition names in dataset,
make sure both order and length of \code{incondition} and \code{outcondition} should match exactly.
}
\examples{
\dontrun{
 ITDRdata <- ms_conditionrename(ITDRdata, incondition=c("staurosporine.3","staurosporine.4"), outcondition=c("ST.1","ST.2"))
}


}
