% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_dataaverage.R
\name{ms_data_average}
\alias{ms_data_average}
\title{ms_data_average}
\usage{
ms_data_average(data, nread = 10, rep = "r", calPSM = FALSE,
  calsd = FALSE)
}
\arguments{
\item{data}{dataset to be filtered}

\item{nread}{number of reading channels or sample treatements, default value
is 10}

\item{rep}{the rep indicator on the naming of experimental conditions}

\item{calPSM}{whether to calculate arithmatic mean of PSMs, Unique peptides
and count numbers}

\item{calsd}{whether to calculate standard deviation on readings, default FALSE}
}
\value{
dataframe containing the averaged readings
}
\description{
Function to perform reading averaging on dataframe
}
\examples{
\dontrun{
data_averaged <- ms_data_average(data_complete_set, rep="")
}


}
