% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_dircreate.R
\name{ms_dircreate}
\alias{ms_dircreate}
\title{ms_dircreate}
\usage{
ms_dircreate(dirname, data)
}
\arguments{
\item{dirname}{a file name to be used for sub-working directory naming}

\item{data}{the dataset to be appended with the sub-working directory name}
}
\value{
a dataframe
}
\description{
Function to create a sub-working directory to store relevant results
}
\keyword{internal}
