% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_dircreate.R
\name{ms_directory}
\alias{ms_directory}
\title{ms_directory}
\usage{
ms_directory(data, dataname)
}
\arguments{
\item{data}{the dataset to be appended with the sub-working directory name}
}
\value{
a working directory
}
\description{
Function to extract the directory information from data or assign a new
directory to data
}
\keyword{internal}
