% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_to_eSet.R
\name{ms_eSet_to_df}
\alias{ms_eSet_to_df}
\title{ms_eSet_to_df}
\usage{
ms_eSet_to_df(expressionset, columnorder = NULL)
}
\arguments{
\item{expressionset}{expression dataset to transform into a dataframe}

\item{columnorder}{a vector for specifying the order of columns}
}
\value{
a dataframe
}
\description{
Function to reformat an expressionset (eSet) to a dataframe format
}
\examples{
\dontrun{
}


}
\keyword{internal}
