% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fileread_filewrite.R
\name{ms_fileread}
\alias{ms_fileread}
\title{ms_fileread}
\usage{
ms_fileread(file)
}
\arguments{
\item{file}{name of the txt file, in the format of "./subfolder/datafile.txt"}
}
\value{
a dataframe
}
\description{
Function to read in dataset from tab delimited files.
The file should be within a subfolder under current working directory
The input such as in the format of "./subfolder/datafile.txt", where the first . indicated the current working directory
}
\examples{
\dontrun{
 ms_fileread("./subfolder/datafile.txt")
}


}
