% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fileread_filewrite.R
\name{ms_filewrite}
\alias{ms_filewrite}
\title{ms_filewrite}
\usage{
ms_filewrite(data, filename, outdir = NULL, withdescription = TRUE,
  autodir = TRUE)
}
\arguments{
\item{data}{dataset to save into txt file}

\item{filename}{name of the txt file}

\item{outdir}{the subfolder name to save the txt file}

\item{withdescription}{whether the dataset contains the description column, default set to TRUE}

\item{autodir}{whether to automatically use the sub-working directory associated to the data, defalut set to TRUE}
}
\description{
Function to write dataset into tab delimited files
}
\examples{
\dontrun{
 ms_filewrite(data, "file_to_mydata.txt")
}


}
