% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fitting.R
\name{ms_fitting}
\alias{ms_fitting}
\title{ms_fitting}
\usage{
ms_fitting(data, nread = 10, topasone = TRUE, halforiginal = TRUE,
  writetofile = TRUE, keepfittedvalue = FALSE)
}
\arguments{
\item{data}{dataset to be fitted}

\item{nread}{number of reading channels or sample treatements,
default value 10}

\item{topasone}{whether the top plateau has to be fixed, i.e., 1.0}

\item{halforiginal}{whether calculate Tm based on the point with absolute
reading of 0.5, default is TRUE; otherwise, calculate Tm based on the
half way between top and bottom plateaus, which is essentially
the inflection point}

\item{writetofile}{whether to keep a local file copy of the original data
with fitting parameters, default set to TRUE}

\item{keepfittedvalue}{whether to keep the fitted data at each temperature
points with fitting parameters, default set to FALSE}
}
\value{
a dataframe
}
\description{
Function to perform sigmoidal melt curve fitting to generate fitting
parameters for each entry/curve
}
\examples{
\dontrun{
LY_fitted <- ms_fitting(LY_scaled)
}


}
