% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_forcescaling.R
\name{ms_forcescaling}
\alias{ms_forcescaling}
\title{ms_force_scaling}
\usage{
ms_forcescaling(data, refcondition = NULL, nread = 10,
  reordertemp = FALSE, writefactortofile = TRUE,
  bottomlabel = "Temperature",
  filename = "CETSA_normalization_factors.txt")
}
\arguments{
\item{data}{dataset to be scaled}

\item{refcondition}{condition names to specify the reference dataset}

\item{nread}{number of reading channels or sample treatements}

\item{reordertemp}{whether to check and re-arrange the treatment temperatures
in ascending order, using the readings from lowest temperature
group as the reference to derive ratios, default set to FALSE since we
typically label the CETSA melt curve samples from low to high temperatures
in the TMT10}

\item{writefactortofile}{whether to save a copy of scaling factors}

\item{bottomlabel}{textual label at the bottom of the plot}

\item{filename}{name for the file}
}
\value{
a dataframe
}
\description{
Function to apply forced systematic scaling to CETSA melt curve dataset, i.e.,
to force scale the median of dataset to be same as the median of a subset of dataset (reference dataset)
}
\examples{
\dontrun{
LY_forcescaled <- ms_forcescaling(LY_cleaned)
}


}
