% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_composite.R
\name{ms_getGene}
\alias{ms_getGene}
\title{ms_getGene}
\usage{
ms_getGene(data)
}
\arguments{
\item{data}{a dataframe with id and description column}
}
\value{
a dataframe with an extra gene column
}
\description{
Functions to add one extra column with gene information from the Uniprot description
}
\examples{
\dontrun{
}

}
