% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_ggplotting_errorbar.R
\name{ms_ggplotting_errorbar}
\alias{ms_ggplotting_errorbar}
\title{ms_ggplotting_errorbar}
\usage{
ms_ggplotting_errorbar(data, legenddata = NULL, levelvector = NULL,
  nread = 10, rep = "r", minireplicate = NULL, fitremout = FALSE,
  bottomcutoff = 0.4, topcutoff = 0.8, topasone = TRUE,
  pfdatabase = FALSE, printBothName = TRUE, printGeneName = FALSE,
  printTm = TRUE, annotypos = 0.5, annotyinterval = 0.08,
  layout = c(5, 5), external = TRUE, presetcolor = TRUE,
  colorpanel = NULL, toplabel = "CETSA data plotting_barplot",
  leftlabel = "Non-denatured protein fraction",
  bottomlabel = "Temperature", pdfname = "ggplotting.pdf",
  pdfheight = 12, pdfwidth = 12)
}
\arguments{
\item{data}{isothermal dataset to plot}

\item{legenddata}{dataset used for condition extraction, at least one protein
inside this dataset should contains the full set of experiment conditions,
most of the time there is no need to supply, just use the data instead}

\item{levelvector}{a vector of experimental conditions, not complusory for isothermal functions}

\item{nread}{number of reading channels or sample treatements}

\item{rep}{the replicate indicator used in the naming of experimental conditions,
such as "r", or "rep" or ""}

\item{minireplicate}{ideally at least three replicates of measurments for the
same proteins are required to plot in a box plot format}

\item{fitremout}{whether to segregate the proteins with messy melt curves}

\item{bottomcutoff}{the average of the last three points should be lower than
specified bottom cutoff value, which is 0.4 by default}

\item{topcutoff}{the average of the first three points should be higher than
specified bottom cutoff value, which is 0.8 by default}

\item{topasone}{whether the top plateau has to be fixed, i.e., 1.0}

\item{printTm}{whether to annotate the plots with Tm values, however in this version,
when there are one samples, the Tm annotation would be switched off}

\item{annotypos}{the starting y-axis position of textual annotation, default value 0.5}

\item{annotyinterval}{the interval on y-axis for textual annotation, default valule 0.08}

\item{layout}{a vector indicating the panel layout for multi-panel plots per page}

\item{presetcolor}{whether to use the pre-defined color scheme}

\item{colorpanel}{a vector of customizable color scheme provided by the user}

\item{toplabel}{textual label at the top part of the page}

\item{leftlabel}{textual label at the left side of the page}

\item{bottomlabel}{textual label at the bottom part of the page}

\item{pdfname}{name for the pdf plots file}

\item{pdfheight}{a number indicate the height of pdf file, default value 12}

\item{pdfwidth}{a number indicate the width of pdf file, default value 12}
}
\value{
a list of ggplot2 object
}
\description{
Function to generate pdf files with multipanel ggplots for replicated
(ideally at least three) CETSA melt curve data, in the format of a mean +/- errorbar
}
\examples{
\dontrun{
ms_ggplotting_errorbar(LY_fitted)
}


}
