% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_dataaverage.R
\name{ms_goodctrlcurve_selection}
\alias{ms_goodctrlcurve_selection}
\title{ms_goodctrlcurve_selection}
\usage{
ms_goodctrlcurve_selection(data, nread = 10, ctrlcond = NULL,
  topCVcutoff = 0.1, bottomCVcutoff = 0.1, topcutoff = NULL,
  bottomcutoff = NULL, nMAD = 2.5)
}
\arguments{
\item{data}{dataset to look for complete set}

\item{nread}{number of reading channels or sample treatements, default value
is 10}

\item{ctrlcond}{by default the function only apply on the data containing condition
keyword "Ctrl"/"DMSO", the user could add customized condition keyword using this argument}

\item{topCVcutoff}{the threshold CV value used to define the top plateau cutoff, default value is 0.1}

\item{bottomCVcutoff}{the threshold CV value used to define the bottom plateau cutoff, default value is 0.1}

\item{topcutoff}{the threshold value to control the minimal top plateau, when provided}

\item{bottomcutoff}{the threshold value to control the maximal bottom plateau, when provided}

\item{nMAD}{when top or bottom CV cutoff were not provided, the cutoff value
would be determined using MAD scheme, nMAD indicates the significance level
of MAD cutoff, default value is 2.5}
}
\value{
dataframe containing the subset of melting curves with good melting profile
}
\description{
Function to perform good quality melting curve selection on dataframe
The main criteria including averaged R2>0.8, with steady plateau
at the top and bottom part of the melting curves
It could be better to use the fitted reading instead of raw readings (TBA)
}
\examples{
\dontrun{
data_good <- ms_goodcurve_selection(data_scaled, bottomcutoff=0.3)
}


}
