% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_IDIT_QC.R
\name{ms_IDIT_QC}
\alias{ms_IDIT_QC}
\title{ms_IDIT_QC}
\usage{
ms_IDIT_QC(data, foldername = NULL, reportname = NULL, nread = 10,
  PSMcutoff = TRUE, PSMthreshold = 3, isdatafitted = TRUE)
}
\arguments{
\item{data}{ITDR or ITTR dataset to perform QC analysis}

\item{foldername}{name for the QC result folder}

\item{reportname}{name for the QC report}

\item{nread}{number of reading channels or sample treatements, default value 10}

\item{PSMcutoff}{whether to apply PSM threshold cutoff on hit selection,
default set to TRUE}

\item{PSMthreshold}{the threshold of averaged PSM numbers to consider protein
quantification as reliable, default value is 3}

\item{isdatafitted}{whether the provided data is associated with
fitting parameters, i.e., after ms_ITDR_fitting or ms_ITTR_fitting}
}
\description{
Function to generate a QC report for the ITDR or ITTR dataset
}
\examples{
\dontrun{
ms_IDIT_QC(ITDRdata_fitted)
ms_IDIT_QC(ITDRdata_scaled, isdatafitted=FALSE, foldername="QC1")
}


}
