% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_innerscale.R
\name{ms_innerplotbox}
\alias{ms_innerplotbox}
\title{ms_innerplotbox}
\usage{
ms_innerplotbox(data, filename, xlabel, ylabel, isratio, isothermalstyle,
  outdir = NULL)
}
\arguments{
\item{data}{dataset to be calculated}

\item{filename}{filename to be specified}

\item{xlabel}{label on x axis to be specified}

\item{ylabel}{label on y axis to be specified}

\item{isratio}{whether the reading data are ratio or abundance}

\item{isothermalstyle}{whether the reading data are from Isothermal or CETSA melt curve data}

\item{outdir}{the subfolder name to save the plot}
}
\description{
Internal function to generate box plot on reading data, according to condition group
}
\examples{
\dontrun{
}


}
\keyword{internal}
