% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_innerscale.R
\name{ms_innerplotmedian}
\alias{ms_innerplotmedian}
\title{ms_innerplotmedian}
\usage{
ms_innerplotmedian(mediandata, filename, xlabel, ylabel, isothermalstyle,
  outdir = NULL)
}
\arguments{
\item{mediandata}{pre-scaling and post-scaling median data, generated from
\code{ms_innerscale} and with a set category factor}

\item{filename}{filename to be specified}

\item{xlabel}{label on x axis to be specified}

\item{ylabel}{label on y axis to be specified}

\item{isothermalstyle}{whether the reading data are from Isothermal or CETSA melt curve data}

\item{outdir}{the subfolder name to save the plot}
}
\description{
Internal function to generate median line plot on reading data,
comparing pre-scaling and post-scaling median change, according to condition group
}
\examples{
\dontrun{
}


}
\keyword{internal}
