% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_innerscale.R
\name{ms_innerscale}
\alias{ms_innerscale}
\title{ms_innerscale}
\usage{
ms_innerscale(data, nread, sumf = FALSE)
}
\arguments{
\item{data}{dataset to be calculated}

\item{nread}{number of reading channels or sample treatements, default value 10}

\item{sumf}{whether to calculate sum or not, default to FALSE, i.e., to calculate median}
}
\value{
a dataframe
}
\description{
Internal function to generate median or sum of reading data, according to condition group
}
\examples{
\dontrun{
}


}
\keyword{internal}
