% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_isoform_cleanup.R
\name{ms_isoform_consolidate}
\alias{ms_isoform_consolidate}
\title{ms_isoform_consolidate}
\usage{
ms_isoform_consolidate(
  data,
  matchtable,
  nread = 10,
  withabd = FALSE,
  weightbycountnum = TRUE
)
}
\arguments{
\item{data}{dataset to be clead-up from isoform ambiguity}

\item{matchtable}{an isoform substitution matching table}

\item{nread}{number of reading channels or sample treatements,
default value 10}

\item{withabd}{whether the dataset contains abundance data as in ITDR/ITTR data,
by default set to FALSE}

\item{weightbycountnum}{by default to consolidate the isoforms}
}
\value{
a dataframe
}
\description{
Function to perform isoform ambiguity cleanup step 2 using the principle of parsimony
when two or more isoforms are identified within the same search result dataset
}
\details{
Note: Data must contain "sumUniPeps","sumPSMs","countNum" columns,
otherwise the mock named columns should be provided.
}
\examples{
\dontrun{
data_cleaned2 <- ms_isoform_consolidate(data_cleaned1,
              matchtable="./subfolder/tobe_consolidated.txt")
data_cleaned2 <- ms_isoform_consolidate(data_cleaned1,
              matchtable="./subfolder/tobe_consolidated.txt", withabd=TRUE)
}


}
