% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_isoform_cleanup.R
\name{ms_isoform_match}
\alias{ms_isoform_match}
\title{ms_isoform_match}
\usage{
ms_isoform_match(data, refdata)
}
\arguments{
\item{data}{dataset to be matched to reference dataset to relieve the problem of isoform ambiguity}

\item{refdata}{dataset used as the check reference, typically the IMPRINTS dataset}
}
\value{
a dataframe
}
\description{
Function to perform an matching of isoform ambiguity between two datasets, ie when
only another isoform of the protein in one dataset is found in the reference dataset,
this isoform is changed to be same as the one present in reference dataset
}
\examples{
\dontrun{
data_meltcurve1 <- ms_isoform_match(data_meltcurve, data_IMPRINTS)
}


}
