% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_isoform_cleanup.R
\name{ms_isoform_resolve}
\alias{ms_isoform_resolve}
\title{ms_isoform_resolve}
\usage{
ms_isoform_resolve(data)
}
\arguments{
\item{data}{dataset to be clead-up from isoform ambiguity}
}
\value{
a dataframe
}
\description{
Function to perform automated isoform ambiguity cleanup using the principle of parsimony
when two or more isoforms are identified among the search result datasets
but for each individual dataset, only one isoform is identified.
}
\examples{
\dontrun{
data_cleaned1 <- ms_isoform_resolve(data_cleaned)
}


}
