% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_isothermal_bar_innerplot.R
\name{ms_isothermal_bar_innerplot}
\alias{ms_isothermal_bar_innerplot}
\title{ms_isothermal_bar_innerplot}
\usage{
ms_isothermal_bar_innerplot(data, legenddata, levelvector, nread,
  minireplicate, witherrorbar, usegradient, fixedy, presetcolor,
  colorpanel, layout, toplabel, leftlabel, bottomlabel)
}
\arguments{
\item{data}{isothermal dataset to plot}

\item{legenddata}{dataset used for condition extraction, at least one protein
inside this dataset should contains the full set of experiment conditions}

\item{levelvector}{a vector of experimental conditions, not complusory for isothermal functions}

\item{nread}{number of reading channels or sample treatements}

\item{minireplicate}{number of replicates to keep in final data, default to NULL}

\item{fixedy}{whether the y-axis should use a fixed range scale}

\item{presetcolor}{whether to use the pre-defined color scheme}

\item{colorpanel}{a vector of customizable color scheme provided by the user}

\item{layout}{a vector indicating the panel layout for multi-panel plots per page}

\item{toplabel}{textual label at the top of the page}

\item{leftlabel}{textual label at the left side of the page}

\item{bottomlabel}{textual label at the bottom of the page}
}
\value{
a list of ggplot2 object
}
\description{
Internal function to generate ggplot objects for isothermal data in a bar plot format
}
\examples{
\dontrun{
}


}
\keyword{internal}
