% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_isothermal_innerplot.R
\name{ms_isothermal_innerplot}
\alias{ms_isothermal_innerplot}
\title{ms_isothermal_innerplot}
\usage{
ms_isothermal_innerplot(data, legenddata, nread, nreplicate, loess, dotconnect,
  PSManno, PSM_annod, Pep_annod, xlinear, xlog10, xsqrt, xinterval, fixedy,
  presetcolor, colorpanel, layout, top_label, bottom_label)
}
\arguments{
\item{data}{isothermal dataset to plot}

\item{legenddata}{dataset used for condition extraction, at least one protein
inside this dataset should contains the full set of experiment conditions}

\item{nread}{number of reading channels or sample treatements}

\item{nreplicate}{number of replicates}

\item{loess}{whether to perform curve fitting using loess model}

\item{dotconnect}{whether to simply dot connect the readings for each curve}

\item{PSManno}{whether to annotate the plots with PSM and uniPeptide number}

\item{PSM_annod}{dataframe containing the PSM number information}

\item{Pep_annod}{dataframe containing the uniPeptide number information}

\item{xlog10}{whether the x-axis should be in log10 scale}

\item{xsqrt}{whether the x-axis should be in square-root transformed scale}

\item{xinterval}{a number indicating the numerical interval for linear x-axis}

\item{fixedy}{whether the y-axis should use a fixed range scale}

\item{presetcolor}{whether to use the pre-defined color scheme}

\item{colorpanel}{a vector of customizable color scheme provided by the user}

\item{layout}{a vector indicating the panel layout for multi-panel plots per page}

\item{xliner}{whether the x-axis should be in linear scale}

\item{toplabel}{textual label at the top of the page}

\item{bottomlabel}{textual label at the bottom of the page}
}
\value{
a list of ggplot2 object
}
\description{
Internal function to generate ggplot objects for isothermal data
}
\examples{
\dontrun{
}


}
\keyword{internal}
