% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_legend.R
\name{ms_isothermal_legend}
\alias{ms_isothermal_legend}
\title{ms_isothermal_legend}
\usage{
ms_isothermal_legend(legenddata, levelvector, nread, nreplicate,
  presetcolor, colorpanel)
}
\arguments{
\item{legenddata}{dataset used for condition extraction, at least one protein
inside this dataset should contains the full set of experiment conditions}

\item{nread}{number of reading channels or sample treatements}

\item{nreplicate}{number of replicates}

\item{presetcolor}{whether to use the pre-defined color scheme}

\item{colorpanel}{a vector of customizable color scheme provided by the user}
}
\value{
a legend object
}
\description{
Internal function to generate legend for isothermal ggplotting
}
\examples{
\dontrun{
}


}
\keyword{internal}
