% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_legend.R
\name{ms_isothermal_legend2}
\alias{ms_isothermal_legend2}
\title{ms_isothermal_legend2}
\usage{
ms_isothermal_legend2(
  legenddata,
  levelvector,
  nread,
  nsample,
  presetcolor,
  colorpanel
)
}
\arguments{
\item{legenddata}{dataset used for condition extraction, at least one protein
inside this dataset should contains the full set of experiment conditions}

\item{nread}{number of reading channels or sample treatements}

\item{nsample}{number of unique experimental samples}

\item{presetcolor}{whether to use the pre-defined color scheme}

\item{colorpanel}{a vector of customizable color scheme provided by the user}
}
\value{
a legend object
}
\description{
Internal function to generate legend for isothermal ggplotting with error bars
}
\examples{
\dontrun{
}


}
\keyword{internal}
