% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_isothermal_scaling.R
\name{ms_isothermal_scaling}
\alias{ms_isothermal_scaling}
\title{ms_isothermal_scaling}
\usage{
ms_isothermal_scaling(data, nread, abdnorm, reftolowest, remloadc, loadcname,
  numcharmix, writefactortofile, bottomlabel, filename)
}
\arguments{
\item{data}{dataset to be scaled}

\item{nread}{number of reading channels or sample treatements}

\item{abdnorm}{whether to apply protein abundance level normalization}

\item{reftolowest}{whether to check and re-arrange the treatment dose
(or time) in ascending order, using the readings from lowest dose (or time)
group as the reference to derive ratios}

\item{remloadc}{whether to remove loading control sample}

\item{loadcname}{the header name of loading control sample}

\item{numcharmix}{whether the treatment names contains both character and
numeric values}

\item{writefactortofile}{whether to save a copy of scaling factors}

\item{bottomlabel}{textual label at the bottom of the plot}

\item{filename}{name for the file}
}
\value{
a dataframe
}
\description{
Internal function to apply systematic scaling to the ITDR or ITTR dataset
}
\examples{
\dontrun{
}


}
\keyword{internal}
