% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_ITDR_filter.R
\name{ms_ITDR_filter}
\alias{ms_ITDR_filter}
\title{ms_ITDR_filter}
\usage{
ms_ITDR_filter(
  data,
  nread = 10,
  checkpointposition = NULL,
  fcthreshold = 0.3,
  R2threshold = 0.8,
  baselineMAD = 0,
  nbaseline = 3,
  nMAD = 2.5,
  checkreplicate = TRUE,
  treatcondition = NULL,
  PSMcutoff = TRUE,
  PSMthreshold = 3
)
}
\arguments{
\item{data}{dataset to be filtered}

\item{nread}{number of reading channels or sample treatements, default value
is 10}

\item{checkpointposition}{refering to the positions of dose points to check
whether their readings surpass fcthreshold, default value is NULL, which
would automatically check the highest 3 dose points}

\item{fcthreshold}{short for fold change threshold, indicate the threshold
fold change compared to baseline (i.e.,1.0) above which the readings can be
considered significantly stabilized or destabilized, default value is 0.3}

\item{R2threshold}{minimal R2 to consider dose-reponse fitting as reliable}

\item{baselineMAD}{MAD of baseline variation, default value is 0; if not
provided, it will be calculated based on the readings from the lowest few
dose points, specified by nbaseline value}

\item{nbaseline}{the number of points (count from lowest concentration) used
for calculate the baseline MAD, default is 3}

\item{nMAD}{the significance level of MAD cutoff, default value is 2.5}

\item{checkreplicate}{whether to check replicated measurements, when set to
TRUE, further segregate the proteins that all the replicated measurements from
at least one experimental condition surpass fold change threshold, default
set to TRUE}

\item{treatcondition}{specify the condition names to perform replicate check,
no need to specify the replicate part of the full condition names, useful when
multiple experimental conditions are contained in the dataset}

\item{PSMcutoff}{whether to apply PSM threshold cutoff on hit selection,
default set to TRUE}

\item{PSMthreshold}{the threshold of averaged PSM numbers to consider protein
quantification as reliable, default value is 3}
}
\value{
a list of segregated dataframes
}
\description{
Function to perform ITDR data segregation and positive hits selection based
on desired fc threshold level and the associated fitting parameters
}
\examples{
\dontrun{
ITDRdata_filtered <- ms_ITDR_filter(ITDRdata_fitted, fcthreshold=0.3,
                     checkreplicate=TRUE, PSMcutoff=TRUE)
}


}
