% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_isothermal_fitting.R
\name{ms_ITDR_fitting}
\alias{ms_ITDR_fitting}
\title{ms_ITDR_fitting}
\usage{
ms_ITDR_fitting(data, nread = 10, fc = 0.3, forcestart = FALSE,
  writetofile = TRUE, keepfittedvalue = FALSE)
}
\arguments{
\item{data}{dataset to be fitted}

\item{nread}{number of reading channels or sample treatements,
default value 10}

\item{fc}{short for fold change, indicate the response level the fitting
function used to back-calculate the corresponding effective concentration}

\item{forcestart}{whether the base of dose-reponse curve start from 1.0,
default set to FALSE to allow free baseline}

\item{writetofile}{whether to keep a local file copy of the original data
with fitting parameters, default set to TRUE}

\item{keepfittedvalue}{whether to keep the fitted data at each dose points
with fitting parameters, default set to FALSE}
}
\value{
a dataframe
}
\description{
Function to perform dose-response function fitting to generate fitting
parameters for each entry/curve
}
\examples{
\dontrun{
ITDRdata_fitted <- ms_ITDR_fitting(ITDRdata_scaled)
}


}
