% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_isothermal_fitting.R
\name{ms_ITDR_fitting_new}
\alias{ms_ITDR_fitting_new}
\title{ms_ITDR_fitting_new}
\usage{
ms_ITDR_fitting_new(data, nread = 10, calMDT = TRUE, modelbestR2 = TRUE,
  fc = 0, nbaseline = 3, baselineMAD = NULL, nMAD = 2.5,
  writetofile = TRUE, keepfittedvalue = FALSE)
}
\arguments{
\item{data}{dataset to be fitted}

\item{nread}{number of reading channels or sample treatements,
default value 10}

\item{calMDT}{whether to calculate the Mininal Dose Threshold(MDT), this is
particularly useful when do follow-up analysis such as R2-AUC plot, since v0.3.7,
default set to TRUE}

\item{modelbestR2}{whether to choose the model with largest R2, otherwise to
choose the model with most representative (most close to median) R2}

\item{fc}{short for fold change, since v0.3.7, default set to 0}

\item{writetofile}{whether to keep a local file copy of the original data
with fitting parameters, default set to TRUE}

\item{keepfittedvalue}{whether to keep the fitted data at each dose points
with fitting parameters, default set to FALSE}
}
\value{
a dataframe
}
\description{
Function to perform dose-response function fitting to generate fitting
parameters for each entry/curve, several models were tried and the best
(or the most representative) one was captured
}
\examples{
\dontrun{
ITDRdata_fitted <- ms_ITDR_fitting(ITDRdata_scaled)
}


}
