% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_ITDR_ggplotting.R
\name{ms_ITDR_ggplotting}
\alias{ms_ITDR_ggplotting}
\title{ms_ITDR_ggplotting}
\usage{
ms_ITDR_ggplotting(data, legenddata = NULL, nread = 10,
  remsinglecondprot = TRUE, PSMcutoff = FALSE, PSMthreshold = 3,
  nreplicate = 2, remfragment = FALSE, remribosomal = FALSE,
  orderEC = FALSE, orderAUC = FALSE, plotseq = NULL, loess = FALSE,
  dotconnect = FALSE, pfdatabase = FALSE, printGeneName = FALSE,
  PSManno = FALSE, unit = "mM", xlinear = FALSE, xlog10 = TRUE,
  xsqrt = FALSE, xcubert = FALSE, xinterval = NULL, fixedy = FALSE,
  layout = c(5, 5), presetcolor = TRUE, colorpanel = NULL,
  top_label = "ITDR CETSA data plotting", pdfname = "ITDR_ggplotting.pdf",
  external = TRUE)
}
\arguments{
\item{data}{ITDR dataset to plot}

\item{legenddata}{dataset used for condition extraction, at least one protein
inside this dataset should contains the full set of experiment conditions,
most of the time there is no need to supply, just use the data instead}

\item{nread}{number of reading channels or sample treatements, default value 10}

\item{remsinglecondprot}{whether orphan proteins to be plotted,
default to exclude them from plotting}

\item{PSMcutoff}{whether to apply PSM threshold cutoff on hit selection,
default set to FALSE}

\item{PSMthreshold}{the threshold of averaged PSM numbers to consider protein
quantification as reliable, default value is 3}

\item{nreplicate}{number of replicates, default value is 2}

\item{orderEC}{whether to order the plots based on EC (Effective concentration)}

\item{orderAUC}{whether to order the plots based on AUC (Area Under the Curve)}

\item{plotseq}{a vector of plots arragement sequence (composite ID)}

\item{loess}{whether to perform curve fitting using loess model}

\item{dotconnect}{whether to simply dot connect the readings for each curve}

\item{PSManno}{whether to annotate the plots with PSM and uniPeptide number}

\item{unit}{textual annotation for the dose unit, default is "mM"}

\item{xlinear}{whether the x-axis should be in linear scale}

\item{xlog10}{whether the x-axis should be in log10 scale}

\item{xsqrt}{whether the x-axis should be in square-root transformed scale}

\item{xcubert}{whether the x-axis should be in cube-root transformed scale}

\item{xinterval}{a number indicating the numerical interval for linear x-axis}

\item{fixedy}{whether the y-axis should use a fixed range scale}

\item{layout}{a vector indicating the panel layout for multi-panel plots
per page, default value is c(5,5)}

\item{presetcolor}{whether to use the pre-defined color scheme}

\item{colorpanel}{a vector of customizable color scheme provided by the user}

\item{toplabel}{textual label at the top of the page}
}
\value{
a list of ggplot2 object
}
\description{
Function to generate pdf files with multipanel ggplots for ITDR data
}
\examples{
\dontrun{
ms_ITDR_ggplotting(ITDRdata_filtered[[1]], orderAUC=TRUE)
}


}
