% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_ITDR_rawread.R
\name{ms_ITDR_rawread}
\alias{ms_ITDR_rawread}
\title{ms_ITDR_rawread}
\usage{
ms_ITDR_rawread(filevector, fchoose = FALSE, dose = c(4, 1, 0.25, 0.0625,
  0.015625, 0.00391, 0.00098, 0.000245, 6e-05, 0), nread = 10,
  abdread = TRUE, PD21 = TRUE, refchannel = "126", channels = c("126",
  "127N", "127C", "128N", "128C", "129N", "129C", "130N", "130C", "131"))
}
\arguments{
\item{filevector}{a file name or a vector of filenems to import}

\item{fchoose}{whether to choose file interactively, default set to FALSE}

\item{dose}{a vector of doses of compound applied to ITDR samples, in the same order as channels}

\item{nread}{number of reading channels, should match the number of channels used, default value 10}

\item{abdread}{whether to read in protein abundance data, default set to TRUE}

\item{PD21}{whether the data is searched using Proteome Discoverer 2.1, default set to TRUE}

\item{refchannel}{names of reference channel used in Proteome Discoverer search, default value 126}

\item{channels}{names of the read-in channels, default value c("126","127N","127C","128N","128C","129N","129C","130N","130C","131")}
}
\value{
a dataframe
}
\description{
Function to parse and read in ITDR data from tab delimited files exported from Proteome Discoverer
}
\examples{
\dontrun{
 ITDRdata <- ms_ITDR_rawread(c("SN1.txt", "SN2.txt", "ST1.txt", "ST2.txt"), PD21=TRUE,
 dose=c(20,5,1.25,0.3123,0.078125,0.019531,0.004883,0.001221,0.000305,0))
}



}
\seealso{
\code{\link{ms_ITTR_rawread}} for time response data
}
