% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_ITTR_filter.R
\name{ms_ITTR_filter}
\alias{ms_ITTR_filter}
\title{ms_ITTR_filter}
\usage{
ms_ITTR_filter(data, nread = 10, fcthreshold = 0.3, R2threshold = 0.8,
  nbaseline = 2, baselineMAD = 0, nMAD = 2.5, checkreplicate = FALSE,
  treatcondition = NULL, allpresence = FALSE, ncheckpoint = 3,
  nreplicate = 2, ncondition = 2, PSMcutoff = FALSE, PSMthreshold = 3,
  remfragment = FALSE, remribosomal = FALSE)
}
\arguments{
\item{data}{dataset to be filtered}

\item{nread}{number of reading channels or sample treatements, default value
is 10}

\item{fcthreshold}{short for fold change threshold, indicate the threshold
fold change compared to baseline (i.e.,1.0) above which the readings can be
considered significantly stabilized or destabilized, default value 0.3}

\item{R2threshold}{minimal R2 level to consider time-reponse fitting as reliable}

\item{nbaseline}{the number of points (count from lowest concentration) used
for calculate the baseline MAD, default is 2}

\item{baselineMAD}{MAD of baseline variation, default value is 0; if not
provided, it will be calculated based on the readings from the lowest two dose groups}

\item{nMAD}{the significance level of MAD cutoff, default value is 2.5}

\item{checkreplicate}{whether to check replicated measurements, when set to
TRUE, make sure the proteins with all the replicated measurements from at least
one condition pass fold change threshold will be segregated, default set to FALSE}

\item{treatcondition}{specify the condition names to perform replicate check,
no need to specify the replicate part of the full condition names, default value is NULL}

\item{allpresence}{whether require the proteins be present (i.e., successfully
quantified) in all the different conditions, default set to FALSE}

\item{ncheckpoint}{the number of dose groups (count down from highest
concentration) to check whether their readings surpass fold change threshold}

\item{nreplicate}{number of replicates}

\item{ncondition}{number of treatment conditions}

\item{PSMcutoff}{whether to apply PSM threshold cutoff on hit selection,
default set to FALSE}

\item{PSMthreshold}{the threshold of averaged PSM numbers to consider protein
quantification as reliable, default value is 3}

\item{remfragment}{whether to remove fragment proteins, default set to FALSE}

\item{remribosomal}{whether to remove ribosomal proteins, default set to FALSE}
}
\value{
a list of segregated dataframes
}
\description{
Function to perform ITTR data segregation and positive hits selection based
on desired fc threshold level and the associated fitting parameters
}
\examples{
\dontrun{
ITTRdata_filtered <- ms_ITTR_filter(ITTRdata_fitted, keeptwice=TRUE,
PSMcutoff=TRUE)
}


}
