% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_ITTR_scaling.R
\name{ms_ITTR_scaling}
\alias{ms_ITTR_scaling}
\title{ms_ITTR_scaling}
\usage{
ms_ITTR_scaling(data, nread = 10, abdnorm = TRUE, reftolowest = TRUE,
  remloadc = FALSE, loadcname = "C", numcharmix = FALSE,
  writefactortofile = TRUE, bottomlabel = "Treatment time",
  filename = "ITTR_normalization_factors.txt")
}
\arguments{
\item{data}{dataset to be scaled}

\item{nread}{number of reading channels or sample treatements, default value 10}

\item{abdnorm}{whether to apply protein abundance level normalization,
default set to TRUE}

\item{reftolowest}{whether to check and re-arrange the treatment dose,
(or time) in ascending order, using the readings from lowest dose (or time)
group as the reference to derive ratios, default set to TRUE}

\item{remloadc}{whether to remove loading control sample, default set to FALSE}

\item{loadcname}{the header name of loading control sample}

\item{numcharmix}{whether the treatment names contains both character and
numeric values}

\item{writefactortofile}{whether to save a copy of scaling factors,
default set to TRUE}

\item{bottomlabel}{textual label at the bottom of the plot}

\item{filename}{name for the file}
}
\value{
scaled dataset in dataframe format
}
\description{
Function to apply systematic scaling to the ITTR dataset
}
\examples{
\dontrun{
ITTRdata_scaled <- ms_ITTR_scaling(ITTRdata_cleaned)
}


}
