% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_melt_curve_summary.R
\name{ms_melt_curve_summary}
\alias{ms_melt_curve_summary}
\title{ms_melt_curve_summary}
\usage{
ms_melt_curve_summary(data, nread = 10, contrast = NULL, writetofile = TRUE)
}
\arguments{
\item{data}{dataset after ms_fitting() function}

\item{nread}{number of reading channels or sample treatements,
default value 10}

\item{contrast}{a character to indicate the contrasting treatment conditions
typically in a format similar as "Drug-Ctrl"}

\item{writetofile}{whether to keep a local file copy of the summary,
default set to TRUE}
}
\value{
a dataframe
}
\description{
Function to summarize the melt curve fitting parameters and
perform statistical analysis for each protein
}
\examples{
\dontrun{
LY_summary <- ms_melt_curve_summary(LY_fitted)
}


}
