% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_legend.R
\name{ms_melt_legend}
\alias{ms_melt_legend}
\title{ms_melt_legend}
\usage{
ms_melt_legend(legenddata, nread, colorpanel)
}
\arguments{
\item{legenddata}{dataset used for condition extraction, at least one protein
inside this dataset should contains the full set of experiment conditions}

\item{nread}{number of reading channels or sample treatements}

\item{colorpanel}{a vector of customizable color scheme provided by the user}
}
\value{
a legend object
}
\description{
Internal function to generate legend for melt curve ggplotting
}
\examples{
\dontrun{
}


}
\keyword{internal}
