% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_dataaverage.R
\name{ms_remove_outlier_replicate}
\alias{ms_remove_outlier_replicate}
\title{ms_remove_outlier_replicate}
\usage{
ms_remove_outlier_replicate(data, nread = 10, minrep = NULL,
  nsignificance = 2, mode = c("SD", "MAD"), abscutoff = NULL,
  remoutlier = TRUE)
}
\arguments{
\item{data}{dataset to clean up from extreme outlier measurement}

\item{nread}{number of reading channels or sample treatements, default value 10}

\item{minrep}{number of minimal replicates required}

\item{nsignificance}{the significance level of assigning extreme outliers, default
value 2, meaning threshold at the level of mean+2*SD of the whole population}

\item{mode}{either SD (Standard Deviation) or MAD (Median Absolute Deviation)}

\item{abscutoff}{the absolute cutoff level of assigning extreme outliers, default
value NULL}

\item{remoutlier}{whether to remove outlier measurements from replicated dataset,
default set to TRUE}
}
\value{
dataframe containing the subset of dataset after removing the outlier measurements
}
\description{
Function to find and remove the outlier measurements from replicated dataset
(with single experimental condition). This function is similar to
function ms_reproducible_replicate(), however the latter one segregates away
all the measurements associated to that protein.
}
\examples{
\dontrun{
data_good_set <- ms_remove_outlier_replicate(data_scaled)
}


}
\keyword{internal}
