% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_dataaverage.R
\name{ms_reproducible_replicate}
\alias{ms_reproducible_replicate}
\title{ms_reproducible_replicate}
\usage{
ms_reproducible_replicate(
  data,
  nread = 10,
  variancecutoff = TRUE,
  nMAD_var = 2.5
)
}
\arguments{
\item{data}{dataset to subset reproducible replicates}

\item{nread}{number of reading channels or sample treatements, default value
is 10}

\item{variancecutoff}{whether to check the variance of the readings
from replicated measurements, default is TRUE}

\item{nMAD_var}{the variance cutoff value would be determined using MAD scheme,
nMAD_var indicates the significance level of MAD cutoff, default value is 2.5}
}
\value{
dataframe containing the reproducible replicates subset
}
\description{
Function to perform subset the reproducible replicated measurements on dataframe
}
\examples{
\dontrun{
data_reproducible_set <- ms_reproducible_replicate(data_complete_set)
}


}
\keyword{internal}
