% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_scaling.R
\name{ms_scaling}
\alias{ms_scaling}
\title{ms_scaling}
\usage{
ms_scaling(data, nread = 10, reorder = FALSE, writefactortofile = TRUE,
  bottomlabel = "Temperature", filename = "CETSA_normalization_factors.txt")
}
\arguments{
\item{data}{dataset to be scaled}

\item{nread}{number of reading channels or sample treatements}

\item{reorder}{whether to check and re-arrange the treatment temperature
in ascending order, using the readings from lowest temperature
group as the reference to derive ratios}

\item{writefactortofile}{whether to save a copy of scaling factors}

\item{bottomlabel}{textual label at the bottom of the plot}

\item{filename}{name for the file}
}
\value{
a dataframe
}
\description{
Function to apply systematic scaling to CETSA melt curve dataset
}
\examples{
\dontrun{
LY_scaled <- ms_scaling(LY_cleaned)
}


}
