% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_dataaverage.R
\name{ms_select_goodmeltcurve}
\alias{ms_select_goodmeltcurve}
\title{ms_select_goodmeltcurve}
\usage{
ms_select_goodmeltcurve(
  data,
  nread = 10,
  ctrlcondition = NULL,
  checkTm = TRUE,
  stableplateau = TRUE,
  topCVcutoff = 0.1,
  bottomCVcutoff = 0.1,
  topcutoff = NULL,
  bottomcutoff = NULL,
  nMAD = 2.5,
  reverse = FALSE
)
}
\arguments{
\item{data}{dataset to look for complete set}

\item{nread}{number of reading channels or sample treatements, default value is 10}

\item{ctrlcondition}{by default the function only apply on the Control treatment data
with specified condition keyword "Ctrl"/"DMSO", the user could add additional customized
condition keyword using this argument}

\item{checkTm}{whether check Tm is valid ie, not NA value}

\item{stableplateau}{whether to select the protein with properly stable plateau in the melt curve,
default set to TRUE}

\item{topCVcutoff}{the threshold CV value used to define the top plateau cutoff, default value is 0.1}

\item{bottomCVcutoff}{the threshold CV value used to define the bottom plateau cutoff, default value is 0.1}

\item{topcutoff}{the threshold value to control the minimal top plateau, when provided}

\item{bottomcutoff}{the threshold value to control the maximal bottom plateau, when provided}

\item{nMAD}{when top or bottom CV cutoff were not provided, the cutoff value
would be determined using MAD scheme, nMAD indicates the significance level
of MAD cutoff, default value is 2.5}

\item{reverse}{whether to return the to-be-removed bad melt curves, default set to FALSE}
}
\value{
dataframe containing the subset of melting curves with good melting profile
}
\description{
Function to perform good quality melting curve selection on dataframe
The main criteria including averaged R2>0.8, with steady plateau
at the top and bottom parts of the melting curves.
It could be better to use the fitted reading instead of raw readings (TBA)
}
\examples{
\dontrun{
data_good <- ms_select_goodmeltcurve(data_scaled, bottomcutoff=0.3)
}


}
\keyword{internal}
