% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_to_eSet.R
\name{ms_to_eSet}
\alias{ms_to_eSet}
\title{ms_to_eSet}
\usage{
ms_to_eSet(data, matrixonly = FALSE, proteininfo = NULL, nread = NULL,
  pdata = NULL, refeset = NULL, logtransform = FALSE)
}
\arguments{
\item{data}{dataset to transform into an expressionset (eSet) format}

\item{matrixonly}{whether the input dataset is a matrix format, default set to FALSE}

\item{proteininfo}{a dataframe containing the correponding protein id-
protein description information, which is only required when the input data is a matrix}

\item{nread}{number of columns with protein abundance information}

\item{pdata}{a dataframe containing the experimental phenotype data information}

\item{refeset}{a reference eset}

\item{logtransform}{whether the numeric data input should be log2 transformed}
}
\value{
a dataframe
}
\description{
Function to reformat dataframe or matrix data into an expressionset (eSet) format
}
\examples{
\dontrun{
}


}
\keyword{internal}
