% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_2D_caldiff.R
\name{ms_2D_caldiff}
\alias{ms_2D_caldiff}
\title{ms_2D_caldiff}
\usage{
ms_2D_caldiff(data, treatmentlevel = NULL)
}
\arguments{
\item{data}{Normalized dataset to calculate the relative protein abundance differences}

\item{treatmentlevel}{a vector of treatment labels, the control should be the first element,
such as c("DMSO","TNFa","AT26533")}
}
\value{
a dataframe
}
\description{
Function to reformat dataframe or matrix data into an expressionset (eSet) format
}
\examples{
\dontrun{
  MOLM <- ms_2D_caldiff(MOLM, treatmentlevel=c("DMSO","TNFa","AT26533"))
}


}
