% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_dataaverage.R
\name{ms_find_replicate}
\alias{ms_find_replicate}
\title{ms_find_replicate}
\usage{
ms_find_replicate(data, nread = 10, keepfullrep = FALSE)
}
\arguments{
\item{data}{dataset to look for complete set}

\item{nread}{number of reading channels or sample treatements, default value
is 10}

\item{keepfullrep}{whether to only keep the proteins with full replicates in
all the experimental conditions, default is FALSE, therefore, to keep the curves
that in at least one condition with full replicates, which is same as `keepcompleterep`}
}
\value{
dataframe containing the full replicated complete dataset
}
\description{
Function to perform complete replicated dataset subsetting on dataframe
}
\examples{
\dontrun{
data_complete_set <- ms_find_replicate(data_scaled)
}


}
