% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_ggplotting_rep.R
\name{ms_ggplotting_rep}
\alias{ms_ggplotting_rep}
\title{ms_ggplotting_rep}
\usage{
ms_ggplotting_rep(data, legenddata = NULL, levelvector = NULL, nread = 10,
  remsinglecondprot = TRUE, PSMcutoff = FALSE, PSMthreshold = 3,
  remfragment = FALSE, remribosomal = FALSE, fitremout = FALSE,
  nreplicate = 2, topasone = TRUE, normTop = TRUE, dotconnect = FALSE,
  pfdatabase = FALSE, printGeneName = FALSE, PSManno = TRUE,
  presetcolor = TRUE, colorpanel = NULL, commonlegend = TRUE,
  layout = c(5, 5), pdfname = "ggplotting.pdf")
}
\arguments{
\item{data}{isothermal dataset to plot}

\item{legenddata}{dataset used for condition extraction, at least one protein
inside this dataset should contains the full set of experiment conditions}

\item{levelvector}{a vector of conditions, preferably starting with Ctrl sets}

\item{nread}{number of reading channels or sample treatements}

\item{remsinglecondprot}{whether orphan proteins to be plotted,
default value is TRUE, so to exclude them from plotting}

\item{PSMcutoff}{whether to apply PSM threshold cutoff on hit selection,
default set to FALSE}

\item{PSMthreshold}{the threshold of averaged PSM numbers to consider protein
quantification as reliable, default value is 3}

\item{fitremout}{whether to segregate the proteins with messy melt curves}

\item{nreplicate}{number of replicates, default value is 2, change the value
accordingly, this is relavent to automatic coloring scheme,
up to 4 is possible for now}

\item{topasone}{whether the top plateau has to be fixed, i.e., 1.0}

\item{normTop}{whether to normalize the AUC based on Top three readings}

\item{dotconnect}{whether to simply dot connect the readings for each curve}

\item{PSManno}{whether to annotate the plots with PSM and uniPeptide number}

\item{presetcolor}{whether to use the pre-defined color scheme}

\item{colorpanel}{a vector of customizable color scheme provided by the user}

\item{commonlegend}{whether to use one common legend for whole page of plots}

\item{layout}{a vector indicating the panel layout for multi-panel plots per page}

\item{pdfname}{name for the pdf plots file}
}
\value{
a list of ggplot2 object
}
\description{
Function to generate pdf files with multipanel ggplots for melt curve data,
particularly for dataset with replicate runs
}
\examples{
\dontrun{
ms_ggplotting_rep(LY_scaled,
  levelvector=c("Ctrl.1", "Ctrl.2", "Treatment.1","Treatment.2"))
}


}
