% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_IDIT_R2AUC.R
\name{ms_IDIT_R2AUC}
\alias{ms_IDIT_R2AUC}
\title{ms_IDIT_R2AUC}
\usage{
ms_IDIT_R2AUC(
  data,
  nread = 10,
  printBothName = FALSE,
  printGeneName = TRUE,
  pfdatabase = FALSE,
  rep = "r",
  onlyshowstabilized = FALSE,
  simpleAUC = TRUE,
  preaveraged = FALSE,
  normalizedAUC = FALSE,
  refkeyword = "37C",
  AUC_uplimit = NULL,
  AUC_lowlimit = NULL,
  log2scale = FALSE,
  nMAD = 2.5,
  baselineMAD = 0,
  fcthreshold = 0,
  checkpointposition = NULL,
  keepreplicate = FALSE,
  keepfullrep = FALSE,
  stableref = FALSE,
  stableref_nMAD = 3.5,
  colorbackground = TRUE,
  graybackground = FALSE,
  nodesizebyMDT = TRUE,
  colornodes = TRUE,
  labelnodes = TRUE,
  yscale = NULL,
  PSMcutoff = TRUE,
  PSMthreshold = 3,
  PSMcutoffbycondition = FALSE,
  idtoshow = NULL,
  idtoexclude = NULL,
  plottitle = NULL,
  returnplot = FALSE
)
}
\arguments{
\item{data}{ITDR or ITTR dataset to generate R2 vs AUC plot, the provided
data should be associated with fitting parameters, i.e., after ms_ITDR_fitting
or ms_ITTR_fitting}

\item{nread}{number of reading channels or sample treatements, default value 10}

\item{printBothName}{whether to print both gene name and protein name, default set to FALSE}

\item{printGeneName}{whether to print only gene name, default set to TRUE, when
both printBothName and printGeneName are FALSE, only the protein name is shown}

\item{pfdatabase}{whether the data is a malaria dataset, default set to FALSE}

\item{rep}{the rep indicator used in the naming of experimental conditions,
such as "r", or "rep" or ""}

\item{onlyshowstabilized}{whether to show only stablized hits, ie,
direct targets, default set to FALSE}

\item{simpleAUC}{whether to perform a simple calculation of AUC, default set to TRUE;
note it is a bit "tricky" to interpret the AUC value in logarithmic space for
most of ITDR data, suggest to just use simpleAUC by default}

\item{preaveraged}{whether the data has already been averaged from replicates,
default set to FALSE}

\item{normalizedAUC}{whether to use the AUC values against a reference control
such as 37C readings, default set to FALSE}

\item{refkeyword}{a keyword used for the indication of reference control, as
most of the time the reference is 37C expression level, default value is 37C}

\item{log2scale}{whether to transform the readings into log2 scale, default set to FALSE}

\item{nMAD}{level of significance of AUC values, default set at 2.5}

\item{baselineMAD}{MAD of baseline variation, default value is 0}

\item{fcthreshold}{short for fold change threshold, similar parameter as in
ms_ITD/TR_filter() function, however this is an added-on selection feature
in this R2AUC plot, and applied to the averaged (and/or normalized) data,
so default value is 0}

\item{checkpointposition}{refering to the positions of dose points to check
whether their readings surpass fcthreshold, default value is NULL, which
would automatically check the highest 3 dose points}

\item{keepreplicate}{whether to only keep the curves that are measured with
all replicates under at least one experimental conditions, default set to
FALSE, when there is only one experimental condition in the dataset, this
means only keep the full replicates, ie, same as keepfullrep=TRUE}

\item{keepfullrep}{whether to only keep the curves that are measured with
all replicates under all experimental conditions, default set to FALSE, this
is useful for the dataset containing more than one experimental condtion}

\item{stableref}{whether to only keep the curves with stable/flat reference
(most of the time, 37C) curves, default set to FALSE}

\item{stableref_nMAD}{level of significance for control of the reference
(most of the time, 37C) sample stability, default set at 3.5}

\item{colorbackground}{whether to color the background with color and shape
according to the condition, default set to TRUE}

\item{graybackground}{whether to use gray color for the background but
different shapes according to the condition, default set to FALSE}

\item{nodesizebyMDT}{whether to set node size according to MDT/MTT value,
ie the EC or ET column, default set to TRUE}

\item{colornodes}{whether to fill the nodes in different colors if possible,
default set to TRUE}

\item{labelnodes}{whether to text-label the nodes, default set to TRUE}

\item{yscale}{a two-element vector to indicate the y-axis scale for plotting
if provided}

\item{PSMcutoff}{whether to apply PSM threshold cutoff on hit selection,
default set to TRUE}

\item{PSMthreshold}{the minimal threshold of averaged PSM numbers to consider
protein quantification as reliable, default value is 3}

\item{PSMcutoffbycondition}{whether to apply PSM threshold cutoff on each
experimental condition, otherwise on all the measured conditions, default set to FALSE}

\item{idtoshow}{a vector containing the uniprot IDs to show on plot
regardless they are hits or not}

\item{idtoexclude}{a vector containing the uniprot IDs to exclude from plot}

\item{plottitle}{a character used to label the title of the plot}

\item{returnplot}{whether to retrieve the plot object, default set to FALSE}
}
\description{
Function to generate a R2 vs AUC plot for the ITDR or ITTR dataset
The R2 (x-axis value), AUC (y-axis) and MDT/MTT (size of dot) are from the input data
}
\examples{
\dontrun{
ms_IDIT_R2AUC(ITDRdata_fitted)
ms_IDIT_R2AUC(ITDRdata_fitted, normalizedAUC=TRUE, stableref=TRUE, yscale=c(0.5,2))
}


}
