% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_innerread.R
\name{ms_innerread}
\alias{ms_innerread}
\title{ms_innerread}
\usage{
ms_innerread(file, fchoose, treatment, nread, abdread, PD21, PD22,
  refchannel, channels)
}
\arguments{
\item{file}{file name to import}

\item{fchoose}{whether to choose file interactively}

\item{treatment}{names of treatments (temperature, dose, time) applied to samples}

\item{nread}{number of reading channels, should match the number of channels used}

\item{abdread}{whether to read in protein abundance data}

\item{PD21}{whether the data is searched using Proteome Discoverer 2.1}

\item{refchannel}{names of reference channel used in Proteome Discoverer search, such as 126}

\item{channels}{names of the read-in channels}

\item{...}{other arguments ignored (for compatibility with generic)}
}
\value{
a dataframe
}
\description{
Internal function to parse and read in data from tab delimited files exported from Proteome Discoverer
}
\examples{
\dontrun{
 ms_innerread("file.txt")
}



}
\keyword{internal}
