% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_isothermal_line_innerplot.R
\name{ms_isothermal_line_innerplot}
\alias{ms_isothermal_line_innerplot}
\title{ms_isothermal_line_innerplot}
\usage{
ms_isothermal_line_innerplot(data, legenddata, levelvector, nread,
  robustfitting, nreplicate, minireplicate, witherrorbar, loess, dotconnect,
  printcount, PSM_annod, Pep_annod, xlinear, xlog10, xsqrt, xcubert, xinterval,
  fixedy, presetcolor, colorpanel, layout, toplabel, leftlabel, bottomlabel,
  shadearea, shadeoutlinecolor, shadefillcolor)
}
\arguments{
\item{data}{isothermal dataset to plot}

\item{legenddata}{dataset used for condition extraction, at least one protein
inside this dataset should contains the full set of experiment conditions}

\item{levelvector}{a vector of experimental conditions, not complusory for isothermal functions}

\item{nread}{number of reading channels or sample treatements}

\item{robustfitting}{whether to apply robust fitting, which is new since v.0.3.7}

\item{nreplicate}{number of replicates}

\item{minireplicate}{number of replicates to keep in final data}

\item{loess}{whether to perform curve fitting using loess model}

\item{dotconnect}{whether to simply dot connect the readings for each curve}

\item{printcount}{whether to annotate the plots with PSM and uniPeptide number}

\item{PSM_annod}{dataframe containing the PSM number information}

\item{Pep_annod}{dataframe containing the uniPeptide number information}

\item{xlog10}{whether the x-axis should be in log10 scale}

\item{xsqrt}{whether the x-axis should be in square-root transformed scale}

\item{xcubert}{whether the x-axis should be in cube-root transformed scale}

\item{xinterval}{a number indicating the numerical interval for linear x-axis}

\item{fixedy}{whether the y-axis should use a fixed range scale}

\item{presetcolor}{whether to use the pre-defined color scheme}

\item{colorpanel}{a vector of customizable color scheme provided by the user}

\item{layout}{a vector indicating the panel layout for multi-panel plots per page}

\item{toplabel}{textual label at the top of the page}

\item{leftlabel}{textual label at the left side of the page}

\item{bottomlabel}{textual label at the bottom of the page}

\item{shadearea}{a four element vector indicating the xmin, xmax, ymin, ymax for x and y axis for shading}

\item{shadeoutlinecolor}{color used to shade the area, default value is black}

\item{shadefillcolor}{color used to shade the area, default value is gray90}

\item{xliner}{whether the x-axis should be in linear scale}
}
\value{
a list of ggplot2 object
}
\description{
Internal function to generate ggplot objects for isothermal data in a curve plot format
}
\examples{
\dontrun{
}


}
\keyword{internal}
