% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_ITTR_fitting.R
\name{ms_ITTR_fitting}
\alias{ms_ITTR_fitting}
\title{ms_ITTR_fitting}
\usage{
ms_ITTR_fitting(data, nread = 10, calMTT = TRUE, fc = 0,
  nbaseline = 3, baselineMAD = NULL, nMAD = 2.5,
  writetofile = TRUE, keepfittedvalue = FALSE)
}
\arguments{
\item{data}{dataset to be fitted}

\item{nread}{number of reading channels or sample treatements,
default value 10}

\item{calMTT}{whether to calculate the Mininal Time Threshold(MTT), this is
useful when do follow-up analysis such as R2-AUC plot}

\item{fc}{short for fold change, indicate the response level the fitting
function used to back-calculate the corresponding effective time,
the set fc value is used only when calMTT=FALSE}

\item{baselineMAD}{MAD of baseline variation, default value is 0; if not
provided, it will be calculated based on the readings from the lowest few
dose points, specified by an integer `nbaseline``}

\item{nMAD}{level of significance, default set at 2.5}

\item{writetofile}{whether to keep a local file copy of the original data
with fitting parameters, default set to TRUE}

\item{keepfittedvalue}{whether to keep the fitted data at each time points
with fitting parameters, default set to FALSE}
}
\value{
a dataframe
}
\description{
Function to perform time-response function fitting to generate fitting
parameters for each entry/curve
}
\examples{
\dontrun{
ITTRdata_fitted <- ms_ITTR_fitting(ITTRdata_scaled)
}


}
