% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_melt_innerplot.R
\name{ms_melt_innerplot}
\alias{ms_melt_innerplot}
\title{ms_melt_innerplot}
\usage{
ms_melt_innerplot(
  data,
  nread,
  topasone,
  dotconnect,
  printcount,
  PSM_annod,
  Pep_annod,
  annotypos,
  annotyinterval,
  colorpanel,
  plotlegend,
  commonlegend,
  toplabel,
  leftlabel,
  bottomlabel,
  withset,
  layout,
  returnplots,
  outdir
)
}
\arguments{
\item{data}{melt curve dataset to plot}

\item{nread}{number of reading channels or sample treatments}

\item{topasone}{whether the top plateau has to be fixed, i.e., 1.0}

\item{dotconnect}{whether to simply dot connect the readings for each curve}

\item{printcount}{whether to annotate the plots with PSM and uniPeptide number}

\item{PSM_annod}{dataframe containing the PSM number information}

\item{Pep_annod}{dataframe containing the uniPeptide number information}

\item{annotypos}{the starting y-axis position of textual annotation}

\item{annotyinterval}{the interval on y-axis for textual annotation}

\item{colorpanel}{a vector of customizable color scheme provided by the user}

\item{plotlegend}{a legend object from ms_melt_legend}

\item{commonlegend}{whether to use one common legend for whole page of plots}

\item{toplabel}{textual label at the top of the page}

\item{leftlabel}{textual label at the left side of the page}

\item{bottomlabel}{textual label at the bottom of the page}

\item{withset}{whether there is set column to perform facet_grid}

\item{layout}{a vector indicating the panel layout for multi-panel plots per page}
}
\value{
a list of ggplot2 object
}
\description{
Internal function to generate ggplot objects for CETSA melt curve data
}
\examples{
\dontrun{
}


}
\keyword{internal}
