% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_rawread.R
\name{ms_rawread}
\alias{ms_rawread}
\title{ms_rawread}
\usage{
ms_rawread(
  filevector,
  fchoose = FALSE,
  temp = c(37, 40, 43, 46, 49, 52, 55, 58, 61, 64),
  nread = 10,
  abdread = FALSE,
  PDversion = 21,
  refchannel = "126",
  channels = NULL
)
}
\arguments{
\item{filevector}{a file name or a vector of filenems to import}

\item{fchoose}{whether to choose file interactively, default set to FALSE}

\item{temp}{a vector of heating temperature applied to CETSA melt samples,
in the same order as channels}

\item{nread}{number of reading channels, should match the number of channels
used, default value 10}

\item{abdread}{whether to read in protein abundance data, default set to FALSE}

\item{PDversion}{which version of Proteome Discoverer the data is searched, possible values 20,21,22,24}

\item{refchannel}{names of reference channel used in Proteome Discoverer
search, default value 126}

\item{channels}{names of the read-in channels, default value NULL, it would
automatically match the provided channel number when it is 10 or 11}
}
\value{
a dataframe
}
\description{
Function to parse and read in CETSA melt curve data from tab delimited files
exported from Proteome Discoverer
}
\examples{
\dontrun{
 LY <- ms_rawread(c("LY_K562_Ctrl_rep1_PD21_Proteins copy.txt",
 "LY_K562_Ctrl_rep2_PD21_Proteins copy.txt",
 "LY_K562_Treatment_rep1_PD21_Proteins copy.txt",
 "LY_K562_Treatment_rep2_PD21_Proteins copy.txt"))
}



}
\seealso{
\code{\link{ms_ITTR_rawread}} for time response data

\code{\link{ms_ITDR_rawread}} for dose response data
}
