% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_scaling.R
\name{ms_scaling}
\alias{ms_scaling}
\title{ms_scaling}
\usage{
ms_scaling(data, nread = 10, reordertemp = FALSE,
  writefactortofile = TRUE, bottomlabel = "Temperature",
  filename = "CETSA_normalization_factors.txt")
}
\arguments{
\item{data}{dataset to be scaled}

\item{nread}{number of reading channels or sample treatments}

\item{reordertemp}{whether to check and re-arrange the treatment temperatures
in ascending order, using the readings from lowest temperature
group as the reference to derive ratios, default set to FALSE since we
typically label the CETSA melt curve samples from low to high temperatures
in the TMT10}

\item{writefactortofile}{whether to save a copy of scaling factors}

\item{bottomlabel}{textual label at the bottom of the plot}

\item{filename}{name for the file}
}
\value{
a dataframe
}
\description{
Function to apply systematic scaling to CETSA melt curve dataset
}
\examples{
\dontrun{
LY_scaled <- ms_scaling(LY_cleaned)
}


}
